refactor/PRI-92 Signup crean up
