+++ vex_app/frontend/src/components/Logout.tsx
@@ -3,16 +3,16 @@ import { authState } from '@/state/authState';
 import { useNavigate } from 'react-router-dom';
 import { useSetRecoilState } from 'recoil';
 
-const Logout = () => {
+export const Logout = () => {
   const navigate = useNavigate();
   const setAuth = useSetRecoilState(authState);
 
   const handleLogout = async () => {
+    // ログアウト処理
     const { error } = await supabase.auth.signOut();
     if (error) {
       alert(`logout error: ${error.message}`);
     } else {
-      alert("logout successful");
       setAuth(false); // 認証状態を更新
       navigate('/login'); // ログインページにリダイレクト
     }
@@ -22,5 +22,3 @@ const Logout = () => {
     <button onClick={handleLogout}>Logout</button>
   );
 };
-
-export default Logout;
