+++ vex_app/frontend/src/pages/Home.tsx
@@ -1,5 +1,5 @@
 import { DatePickerForm } from "@/components/DatePickForm";
-import Logout from "@/components/Logout";
+import { Logout } from "@/components/Logout";
 import useEventData from "@/hooks/useEventData";
 import { useGoogleMap } from "@/hooks/useGoogleMap";
 import { Loader } from '@googlemaps/js-api-loader';
